Extract linked evidence pack validator
